requirements.txt
chromadb
flask
python-dotenv
google-generativeai
sentence-transformers
langchain
langchain-community
pypdf
pdfplumber
gunicorn
psycopg2-binary